Reorganized files; upgraded project to VS2015
When I first posted this project on Github, I was pretty lazy and just
threw all the files in the same directory.  Now I've made an effort to
make the organization more explicit, so the overall structure should be
easier to understand.

I also upgraded the project files to VS2015. After I had done that work
I found out that VS2017 is already available, so I'll upgrade to that
in a separate commit.  Maybe this VS2015 commit will be useful to
people who haven't upgraded yet or something.

I also created a project for the building of the HTML Help file, which
is a pretty obsolete format, but it has search functionality, so I'm
hesitant to get rid of it entirely until I implement searching for the
docs on my web site.

I also created an installer (finally!), so I'm making this a release
just for that. I'm not sure why I never made one. That was dumb of me!